Added smooth rotation and speed acceleration
* It either works or it doesn't
* Needs testing
* There are some unstable things so this is all very unstable
* Do something with this recoil ;)

+++ src/System/Weapon.cpp
@@ -34,10 +34,10 @@ HOA::AttackType Weapon::attackType() const
 	return base_->attackType();
 }
 
-int Weapon::maxSpeed() const
+qreal Weapon::maxSpeed() const
 {
 	//TODO fix maxSpeed/hitRatio conflict
-	return 1;
+	return 1.0;
 }
 
 int Weapon::damage() const
